feat: 完成标准的 main 函数初始化过程

+++ src/file/fileManager.h
@@ -74,7 +74,7 @@ bool FileManager::initStandardJudge() {
     Job *getStandardJudgeList = new SocketWork(&standardJudgeRequest, &standardJudgeCallback, &cm);
     sessionPool->submit(getStandardJudgeList);
     cm.wait();
-    int len = standardJudgeNameList.size();
+    int len = (int) standardJudgeNameList.size();
 
     /// endregion
